refactor(ImageCreator): 修复文本绘制位置的浮点数值问题
1.  将marginTop和marginLeft的计算结果强制转为整数，避免浮点坐标导致的绘制异常
2.  新增完整的单元测试覆盖图片创建流程
3.  新增API响应和不同验证码样式的测试用例
4.  添加了方正雅园纤细简体字体文件资源

+++ src/Image/ImageCreator.php
@@ -116,14 +116,14 @@ class ImageCreator
         $width = $image->width();
         $height = $image->height();
         $padding = $config['textLeftPadding'] ?? 4;
-        $marginTop = $config['marginTop'] ?? ($height / $length);
+        $marginTop = $config['marginTop'] ?? (int) ($height / $length);
         $angle = $config['angle'] ?? 15;
 
         foreach ($text as $key => $char) {
             $marginLeft = $padding + ($key * ($width - $padding) / $length);
             $isChinese = $this->containsChinese($char);
 
-            $image->text($char, $marginLeft, $marginTop, function (FontFactory $font) use ($config, $height, $angle, $isChinese) {
+            $image->text($char, (int) $marginLeft, (int) $marginTop, function (FontFactory $font) use ($config, $height, $angle, $isChinese) {
                 $font->filepath($this->getRandomFont($isChinese));
                 $font->size($isChinese ? random_int($height - 20, $height - 6) : random_int($height - 10, $height));
                 $font->color($this->getRandomColor($config['fontColors'] ?? []));
